project/requirements.txt
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pytest>=7.4.0
joblib>=1.3.0
pyyaml>=6.0
